requirements.txt
# requirements.txt
sentence-transformers
faiss-cpu
python-docx
python-pptx
pymupdf                  # Superior PDF text extraction (replaces pypdf)
ollama
streamlit
streamlit-chat
torch                    # Required by sentence-transformers (often needed explicitly)